% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphModplot.R
\name{graphModplot}
\alias{graphModplot}
\title{Graph Plot Editor}
\usage{
graphModplot(
  x,
  col = "grey80",
  bg = c("red", "black", "blue"),
  pch = 21L,
  length = 0.05,
  pt.cex = 0.75,
  ...
)
}
\arguments{
\item{x}{A \code{\link{graph-class}} object.}

\item{col}{Color of the arrows representing the edges. Default is
\code{"grey80"}.}

\item{bg}{Background colors for the origin, intermediate, and terminal
vertices, respectively. Default is \code{c("red","black","blue")}.}

\item{pch}{Plotting 'character' (see \code{\link[graphics]{points}}) for the
details. Default is \code{21} (a solid dot with a colored background).}

\item{length}{Length of the edge of the arrow head (see
\code{\link[graphics]{arrows}}). The default value is \code{0.05}.}

\item{pt.cex}{The relative point size used for the vertex markers. The
default value is \code{0.75}.}

\item{...}{Additional parameters to be passed to other functions of methods.}
}
\value{
A point geometry object.
}
\description{
Interactively modify the plot of a graph object by manually
changing the location where vertices are being placed.
}
\details{
After calling the function, a mouse-click on the display will
select the nearest vertex. Which one has been selected is indicated by the
background of its marker turning white, with its incoming edges appearing as
red dashed segments and its outgoing edges appearing as blue dashed segments.
A second mouse-click will cause the vertex to be relocated at the location
of the mouse pointer.
}
\examples{
## Create an exemplary graph:
data.frame(
  species = rep(TRUE,13),
  type = c(2,2,3,1,2,2,2,2,2,2,3,3,3),
  x = c(1,3,4,0,1.67,4,1,1.33,2.33,3.33,4.33,4,5),
  y = c(1,1,1,0,0.5,0,-1,0,0,-0.5,-1,-0.5,-0.5),
  row.names = sprintf("V\%d",1:13)
) \%>\%
  st_as_sf(
    coords=c("x","y"),
    crs = NA
  ) \%>\%
  graph \%>\%
  add.edge(
    from = c(1,2,1,5,4,4,5,9,4,8,9,4,7,7,6,6,9,10,10),
    to = c(2,3,5,2,1,5,9,2,8,9,6,7,8,9,3,13,10,12,11),
    data = data.frame(
      distance = c(4.2,4.7,3.9,3.0,3.6,2.7,4.4,3.4,3.6,3.3,
                   4.8,3.2,3.5,4.4,2.5,3.4,4.3,3.1,2.2),
      row.names = sprintf("E\%d",1:19)
    )
  ) -> x

## Original coordinates:
plot(x)

## Edit the vertex locations manually:
x <- graphModplot(x)

## Plot with the new coordinates
plot(x)


}
\seealso{
\code{\link{graph-class}}.
}
\author{
\packageAuthor{RPEM}
Maintainer: \packageMaintainer{RPEM}
}
