% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmean.R
\name{hmean}
\alias{hmean}
\alias{weighted.hmean}
\title{Harmonic Mean}
\usage{
hmean(x)

weighted.hmean(x, w)
}
\arguments{
\item{x}{A numeric vector containing a set of observations.}

\item{w}{A numeric vector containing the observation weights.}
}
\value{
A numeric value.
}
\description{
Functions to calculate the harmonic mean and weighted harmonic mean of a set
of data.
}
\details{
The (weighted) harmonic mean of a data set is the inverse of the
(weighted) mean of the inverse values. As such, any 0 value in the data (or
in the non-zero weighted data in the weighted case) will result in the
functions returning a value of 0. It is worth recall that whereas the value
of the arithmetic mean is strongly influenced by even a few extreme values,
the value of the harmonic mean is strongly influenced by even a few small
values. The function take any numeric values. However, the harmonic mean of
data containing both positive and negative values is unlikely to be stable.
}
\section{Functions}{
\itemize{
\item \code{hmean()}: Harmonic Mean.

Calculates the (unweighted) harmonic mean.

\item \code{weighted.hmean()}: Weighted Harmonic Mean

Calculates the weighted harmonic mean.

}}
\author{
\packageAuthor{RPEM}
}
