% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceMatrix.R
\docType{class}
\name{InflMat-class}
\alias{InflMat-class}
\alias{InflMat}
\alias{print.InflMat}
\alias{nedge.InflMat}
\alias{edge.InflMat}
\alias{edge<-.InflMat}
\alias{edgenames.InflMat}
\alias{edgenames<-.InflMat}
\title{Influence Matrix}
\usage{
InflMat(x)

\method{print}{InflMat}(x, ...)

\method{nedge}{InflMat}(x)

\method{edge}{InflMat}(x)

\method{edge}{InflMat}(x) <- value

\method{edgenames}{InflMat}(x)

\method{edgenames}{InflMat}(x) <- value
}
\arguments{
\item{x}{A \code{\link{graph-class}} or \code{InflMat-class} object.}

\item{...}{Further arguments to be passed internally to other functions or
methods.}

\item{value}{A vector or \code{\link{data.frame}} containing the values to be
given to the \code{InflMat-class} object.}
}
\value{
The returned value depends on the function:
\describe{
\item{InflMat}{A binary influence matrix of the graph with as many rows as
its number of vertices and as many columns as its number of edges.}
}
}
\description{
Functions and methods calculate and manipulate graph influence
matrix.
}
\section{Functions}{
\itemize{
\item \code{InflMat()}: Influence Matrix

Calculates the influence matrix of a phylogenetic graph. The influence matrix
is a binary matrix whose rows and columns correspond to the vertices and
edges of the phylogenetic graph, respectively, and whose elements describe
whether a given edge had been taken by any ancestors of a vertex
(representing extinct of extant species) during evolution (value = 1) or not
(value = 0).

\item \code{print(InflMat)}: Print Graph

A print method for \code{InflMat-class} objects.

\item \code{nedge(InflMat)}: Number of Edges

Get the number of edges in an \code{InflMat-class} object.

\item \code{edge(InflMat)}: Edge Extraction

Extracts the edges of an \code{InflMat-class} object.

\item \code{edge(InflMat) <- value}: Edge Assignment

Assigns edges to an \code{InflMat-class} object.

\item \code{edgenames(InflMat)}: Edge Names Extraction

Extracts the edge names of an \code{InflMat-class} object.

\item \code{edgenames(InflMat) <- value}: Edge Names Assignment

Assigns edge names to an \code{InflMat-class} object.

}}
\examples{
 ## Exemplary graph:
data.frame(
  species = as.logical(c(0,0,1,0,0,0,0,0,0,0,1,1,1)),
  type = c(2,2,3,1,2,2,2,2,2,2,3,3,3),
  x = c(1,3,4,0,1.67,4,1,1.33,2.33,3.33,4.33,4,5),
  y = c(1,1,1,0,0.5,0,-1,0,0,-0.5,-1,-0.5,-0.5),
  row.names = sprintf("V\%d",1:13)
) \%>\%
  st_as_sf(
    coords=c("x","y"),
    crs = NA
  ) \%>\%
  graph \%>\%
  add.edge(
    from = c(1,2,1,5,4,4,5,9,4,8,9,4,7,7,6,6,9,10,10),
    to = c(2,3,5,2,1,5,9,2,8,9,6,7,8,9,3,13,10,12,11),
    data = data.frame(
      distance = c(4.2,4.7,3.9,3.0,3.6,2.7,4.4,3.4,3.6,3.3,
                   4.8,3.2,3.5,4.4,2.5,3.4,4.3,3.1,2.2),
      row.names = sprintf("E\%d",1:19)
    )
  ) -> x

## Calculation of the influence matrix:
y <- InflMat(x)

## Obtain the number of edges:
nedge(y)

## Obtain the edge names:
edgenames(y)

## Obtain the edge data frame:
edge(y)

}
\references{
Guénard, G., Legendre, P., and Peres-Neto, P. 2013. Phylogenetic eigenvector
maps: a framework to model and predict species traits. Methods in Ecology 
and Evolution. 4: 1120--1131

Makarenkov, V., Legendre, L. & Desdevise, Y. 2004. Modelling phylogenetic
relationships using reticulated networks. Zoologica Scripta 33: 89--96

Blanchet, F. G., Legendre, P. & Borcard, D. 2008. Modelling directional
spatial processes in ecological data. Ecological Modelling 215: 325--336
}
\seealso{
\code{\link{PEM-class}} \code{\link{PEM-functions}}
}
\author{
\packageAuthor{RPEM} --
Maintainer: \packageMaintainer{RPEM}
}
