% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-utils.R
\name{lm-utils}
\alias{lm-utils}
\alias{utils}
\alias{model.data}
\alias{Psquare}
\title{Linear Modelling Utility Functions}
\usage{
model.data(
  formula,
  data,
  ...,
  na.action = na.pass,
  contrasts = NULL,
  discard.intercept = TRUE
)

Psquare(obs, prd)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}" (or one that can be
coerced to that class): a symbolic description of the model data to be
prepared. See ‘Details’ in \code{\link{lm}} for further information about
model specification.}

\item{data}{An optional data frame, list or environment (or object coercible
by \code{\link{as.data.frame}} to a data frame) containing the variables. If
not found in data, the variables are taken from environment(formula),
typically the environment from which the function is called.}

\item{...}{Additional parameters to be passed to the method.}

\item{na.action}{A function (of the name of a function) for treating missing
values (\code{NA}s) (default: \code{na.pass}).}

\item{contrasts}{An optional list. See the \code{contrasts.arg} of
\code{\link{model.matrix.default}}. (default: \code{NULL}).}

\item{discard.intercept}{A \code{\link{logical}}; whether of not to discard
the intercept from the model matrix (default: \code{TRUE}).}

\item{obs}{A numeric vector of observations.}

\item{prd}{A numeric vector of model predictions.}
}
\value{
\describe{
  \item{model.data}{A three-member list with member \code{$y} (a vector or
    matrix of response traits), member \code{$x} (a matrix auxiliary traits
    coded as numeric values), and member \code{$terms} (A model
    description).}
  \item{Psquare}{A numeric value.}
}
}
\description{
Utility functions to build linear models using Phylogenetic
Eigenvector Maps among their explanatory variables.
}
\details{
Function \code{model.data} is useful to prepare data to be given as
response and auxiliary trait(s) to other functions such as
\code{\link{evolution.model.PEM}}. In general, the implicit constant term
(intercept) is not useful and can be explicitly discarded.
}
\section{Functions}{
\itemize{
\item \code{model.data()}: Model Data Preparation

Transforms data from various types into a list containing the response
trait(s) and a strictly numeric auxiliary traits data matrix.

\item \code{Psquare()}: Coefficient of Prediction

Calculates the prediction coefficient between observations and model
predictions.

}}
\references{
Guénard, G., Legendre, P., and Peres-Neto, P. 2013. Phylogenetic eigenvector
maps: a framework to model and predict species traits. Methods in Ecology 
and Evolution 4: 1120-1131
}
\author{
\packageAuthor{RPEM}
Maintainer: \packageMaintainer{RPEM}
}
