% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEM-functions.R
\name{PEM-functions}
\alias{PEM-functions}
\alias{PEMweights}
\title{Phylogenetic Eigenvector Maps}
\usage{
PEMweights(d, a = 0, psi = 1)
}
\arguments{
\item{d}{A numeric vector of the evolutionary distances (\code{PEMweights})
or a character string specifying the row name of the edge table (located in
the \code{\link{graph-class}} object's attribute \code{edge}) where the
evolutionary distances (edge lengths) can be found.}

\item{a}{The steepness parameter describing whether changes occur, on
average: progressively long edges (a close to 0) or abruptly at vertices (a
close to 1; default: \code{0}).}

\item{psi}{Relative evolution rate along the edges. This parameter only
becomes relevant when multiple values are assigned to different portions of
the phylogeny (default: \code{1}).}
}
\value{
The returned value depends on the function:
\describe{
  \item{PEMweights}{A set of numeric value to be used as weights during PEM
    calculation.}
}
}
\description{
Functions to calculate and manipulate Phylogenetic Eigenvector
Maps (PEM), which are sets of eigenfunctions describing the structure of a 
phylogenetic graph. Each computation function is briefly described in 
section \code{Functions} below.
}
\section{Functions}{
\itemize{
\item \code{PEMweights()}: PEM Weighting

A power function to obtain the edge weights used during PEM calculation.

}}
\references{
Guénard, G., Legendre, P., and Peres-Neto, P. 2013. Phylogenetic eigenvector
maps: a framework to model and predict species traits. Methods in Ecology 
and Evolution. 4: 1120--1131

Makarenkov, V., Legendre, L. & Desdevise, Y. 2004. Modelling phylogenetic
relationships using reticulated networks. Zoologica Scripta 33: 89--96

Blanchet, F. G., Legendre, P. & Borcard, D. 2008. Modelling directional
spatial processes in ecological data. Ecological Modelling 215: 325--336
}
\seealso{
\code{\link{PEM-class}}
}
\author{
\packageAuthor{RPEM} --
Maintainer: \packageMaintainer{RPEM}
}
