% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemlm-class.R
\docType{class}
\name{pemlm-class}
\alias{pemlm-class}
\alias{pemlm}
\alias{print.pemlm}
\alias{summary.pemlm}
\alias{anova.pemlm}
\alias{predict.pemlm}
\title{PEM Linear Model Interface}
\format{
A \code{pemlm-class} object contains:
\describe{
  \item{auxModel}{A \code{\link{lm}}-class object.}
  \item{resetPEM}{A function ...}
  \item{getPEM}{A function ...}
  \item{aic}{A function ...}
  \item{getIncluded}{A function ...}
  \item{getCandidate}{A function ...}
  \item{scanCandidate}{A function ...}
  \item{forward}{A function ...}
  \item{promote}{A function ...}
  \item{pemModel}{A function ...}
}
}
\usage{
pemlm(formula, data, pem, ..., contrasts = NULL)

\method{print}{pemlm}(x, ...)

\method{summary}{pemlm}(object, ...)

\method{anova}{pemlm}(object, ...)

\method{predict}{pemlm}(
  object,
  newdata = data.frame(row.names = rownames(newloc)),
  newloc,
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)
}
\arguments{
\item{formula}{An object of class "\code{\link{formula}}" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
(see `details` in \code{\link{lm}} for further details on model
specification.}

\item{data}{An optional data frame, list or environment (or object coercible
by \code{\link{as.data.frame}} to a data frame) containing the variables in
the model. If not found in data, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{pemlm} is called.}

\item{pem}{A \code{\link{PEM-class}} object.}

\item{...}{Additional arguments to be passed to the low level regression
fitting functions.}

\item{contrasts}{An optional list. See the \code{contrasts.arg} of
\code{\link{model.matrix.default}}.}

\item{x, object}{A \code{pemlm-class} object.}

\item{newdata}{An optional data frame containing auxiliary traits used for
making predictions.}

\item{newloc}{A data frame containing the graph locations of species targeted
for making predictions.}

\item{se.fit}{A \code{\link{logical}} specifying whether to return standard
errors (default: \code{FALSE}).}

\item{interval}{One of \code{"none"}, \code{"confidence"}, or
\code{"prediction")} specifying the type of interval associated with the
predictions.}

\item{level}{Tolerance/confidence level (default: \code{0.95}).}
}
\value{
\describe{
  \item{pemlm}{A \code{pemlm-class} object.}
  \item{print.pemlm}{\code{NULL} (invisibly).}
  \item{summary.pemlm}{A \code{summary.lm-class} object.}
  \item{anova.pemlm}{An \code{anova-class} data frame.}
  \item{predict.pemlm}{...}
}
}
\description{
An interface to estimate a linear model
}
\details{
...
}
\section{Functions}{
\itemize{
\item \code{pemlm()}: PEM Linear Model

Calculate a PEM-based linear model for estimating trait values.

\item \code{print(pemlm)}: Print PEM Linear Model

A print method for pemlm-class objects.

\item \code{summary(pemlm)}: Summarize pemlm Model

A summary method for \code{pemlm-class} objects.

\item \code{anova(pemlm)}: Anova of a pemlm Model

An anova method for \code{pemlm-class} objects.

\item \code{predict(pemlm)}: Predict Trait Values

A predict method for \code{pemlm-class} objects.

}}
\examples{
## Example here...

}
\references{
Guénard, G., Legendre, P., and Peres-Neto, P. 2013. Phylogenetic eigenvector
maps: a framework to model and predict species traits. Methods in Ecology 
and Evolution 4: 1120-1131
}
\author{
\packageAuthor{RPEM}
Maintainer: \packageMaintainer{RPEM}
}
