% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.dist.R
\name{write.dist}
\alias{write.dist}
\title{Write Distances to a PHYLIP Text File}
\usage{
write.dist(x, file, ...)
}
\arguments{
\item{x}{A distance matrix in the form of a \code{\link{dist}}-class object.}

\item{file}{A character string describing the connection (see the description
of argument \code{description} in function \code{\link{file}} help page).}

\item{...}{Further arguments to be internally passed to function
\code{\link{file}}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
A function to write a \code{\link{dist}}-class object into a
text file.
}
\author{
\packageAuthor{RPEM}
}
