% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{Write Sequences to a FASTA Text File}
\usage{
write.fasta(x, file, linebreak = NULL, sep = "\\n")
}
\arguments{
\item{x}{A \code{\link{character}} vector storing the nucleotide sequences.}

\item{file}{A character string describing the connection (see the description
of argument \code{description} in function \code{\link{file}} help page).}

\item{linebreak}{Interval for line breaking to improve file readability
(default: \code{NULL}, meaning no line breaking).}

\item{sep}{Character used for line breaking (default: \code{"\n"}).}
}
\value{
\code{NULL} (invisibly).
}
\description{
A function to write a vector of strings representing nucleotide
sequences into a text file, with optional line breaks
}
\examples{
## Define a raw vector for storing nuceotide values:
c(Sequence_1 = "ATCG-TTTCG--CCCCA--TTA--TTAA-GTAA-GTAATCTTTCA",
  Sequence_2 = "TTGGCTTCC--TC--CAT-TTC--TTCA-GT-ACG-ATTCTTTTA",
  Sequence_3 = "TTCG-TACC-T-T---A-ATAA--T-AA-GTCTTGTAATCGTTCA") \%>\%
  sapply(charToRaw) \%>\%
  t -> sqn

## Display the raw sequence:
sqn

## Transforming the sequence to character strings
tmp <- concatenate(sqn)
tmp

## Transforming the sequence to character strings without the gaps:
concatenate(sqn, discard="-")

## Write the sequences into a text file:
write.fasta(tmp, file="Sequences.fst", linebreak=15)

## Clean-up:
file.remove("Sequences.fst")
rm(sqn, tmp)

}
\author{
\packageAuthor{RPEM}
}
